Move ProxyRegistry to camelot.core.item_model.proxy_registry
