+++ camelot/core/item_model/__init__.py
@@ -38,6 +38,7 @@ from ..qt import Qt
 from .list_proxy import ListModelProxy
 from .proxy import AbstractModelProxy, AbstractModelFilter
 from .query_proxy import QueryModelProxy
+from .proxy_registry import ProxyRegistry
 
 #
 # Custom Roles
@@ -65,5 +66,6 @@ __all__ = [
     ListModelProxy.__name__,
     ProxyDict.__name__,
     QueryModelProxy.__name__,
+    ProxyRegistry.__name__,
 ]
